feat: finalize GET endpoint for project
